handler functions modification

+++ util/dbHandler.py
@@ -5,10 +5,10 @@ import os
 DB_PATH = os.path.join(os.path.dirname(settings.__file__), settings.DB_NAME)
 CONN = sqlite3.connect(DB_PATH)
 
-def create_table(tableName, title, lines=None):
+def create_table(tableName, title, lines=None, drop=False):
     cursor = CONN.cursor()
     cursor.execute('SELECT name FROM sqlite_master WHERE type="table" AND name="%s"' % tableName)
-    if cursor.fetchone():
+    if cursor.fetchone() and drop:
         cursor.execute('drop table %s' % tableName)
         CONN.commit()
     if not isinstance(title, list):
@@ -18,6 +18,19 @@ def create_table(tableName, title, lines=None):
     if lines:
         cursor.executemany('INSERT INTO %s VALUES (%s)' % (tableName, ','.join(['?' for t in title])), lines)
         CONN.commit()
+    cursor.close()
+
+def insert_line(tableName, line):
+    cursor = CONN.cursor()
+    q = 'INSERT INTO %s VALUES (%s)' % (tableName, ','.join([str(val) for val in line]))
+    cursor.execute(q)
+    CONN.commit()
+    cursor.close()
+
+def insert_lines(tableName, lines):
+    cursor = CONN.cursor()
+    cursor.executemany('INSERT INTO %s VALUES (%s)' % (tableName, ','.join(['?' for t in title])), lines)
+    CONN.commit()
     cursor.close()
 
 def delete_table(tableName):
@@ -33,3 +46,35 @@ def read_table(tableName):
     rec = [line for line in lines]
     cursor.close()
     return rec
+
+def table_exist(tableName):
+    cursor = CONN.cursor()
+    q = 'SELECT name FROM sqlite_master WHERE type="table" AND name="%s"' % tableName
+    cursor.execute(q)
+    return bool(cursor.fetchall())
+
+def get_columns(tableName='userlist', fields=['user_id'], distinct=False):
+    cursor = CONN.cursor()
+    fields_str = ','.join(fields)
+    if distinct:
+        q = 'select distinct %s from %s' % (fields_str, tableName)
+    else:
+        q = 'select %s from %s' % (fields_str, tableName)
+    cursor.execute(q)
+    lines = [line for line in cursor.fetchall()]
+    cursor.close()
+    return lines
+
+def get_all_line_by_fields(tbName='userlist', dict={}, distinct=False):
+    cursor = CONN.cursor()
+    if not distinct:
+        q = 'select * from %s ' % tbName
+    else:
+        q = 'select distinct * from %s ' % tbName
+    if dict:
+        q += 'where '
+        q += ' and '.join(['%s="%s"' % (key, dict[key]) for key in dict])
+    cursor.execute(q)
+    lines = [line for line in cursor.fetchall()]
+    cursor.close()
+    return lines
